[mod/#35] 코드리뷰 반영
