+++ app/src/main/java/sopt/org/millie/data/datasource/SearchDataSource.kt
@@ -1,8 +1,10 @@
 package sopt.org.millie.data.datasource
 
 import sopt.org.millie.core.network.BaseResponse
+import sopt.org.millie.data.dto.response.BookDetailInformationResponseDto
 import sopt.org.millie.data.dto.response.CategoryResponseDto
 
 interface SearchDataSource {
     suspend fun getCategories(): BaseResponse<List<CategoryResponseDto>>
+    suspend fun getBookDetailInformation(bookId: Long): BaseResponse<BookDetailInformationResponseDto>
 }
