+++ app/src/main/java/sopt/org/millie/data/di/DataSourceModule.kt
@@ -4,9 +4,7 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import sopt.org.millie.data.datasource.BookDetailDataSource
 import sopt.org.millie.data.datasource.SearchDataSource
-import sopt.org.millie.data.datasourceimpl.BookDetailDataSourceImpl
 import sopt.org.millie.data.datasourceimpl.SearchDataSourceImpl
 import javax.inject.Singleton
 
@@ -18,10 +16,4 @@ abstract class DataSourceModule {
     abstract fun bindSearchDataSource(
         searchDataSourceImpl: SearchDataSourceImpl,
     ): SearchDataSource
-
-    @Binds
-    @Singleton
-    abstract fun bindBookDetailDataSource(
-        bookDetailDataSourceImpl: BookDetailDataSourceImpl,
-    ): BookDetailDataSource
 }
