+++ app/src/main/java/sopt/org/millie/data/di/ServiceModule.kt
@@ -5,7 +5,6 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import retrofit2.Retrofit
-import sopt.org.millie.data.service.BookDetailService
 import sopt.org.millie.data.service.SearchService
 import javax.inject.Singleton
 
@@ -15,8 +14,4 @@ object ServiceModule {
     @Provides
     @Singleton
     fun provideSearchService(retrofit: Retrofit): SearchService = retrofit.create(SearchService::class.java)
-
-    @Provides
-    @Singleton
-    fun provideBookDetailService(retrofit: Retrofit): BookDetailService = retrofit.create(BookDetailService::class.java)
 }
