--- app/src/main/java/sopt/org/millie/data/repository/BookDetailRepository.kt
@@ -1,7 +1,0 @@
-package sopt.org.millie.data.repository
-
-import sopt.org.millie.data.model.BookDetailModel
-
-interface BookDetailRepository {
-    suspend fun getBookDetailInformation(bookId: Long): Result<BookDetailModel>
-}
