+++ app/src/main/java/sopt/org/millie/data/service/SearchService.kt
@@ -1,10 +1,17 @@
 package sopt.org.millie.data.service
 
 import retrofit2.http.GET
+import retrofit2.http.Path
 import sopt.org.millie.core.network.BaseResponse
+import sopt.org.millie.data.dto.response.BookDetailInformationResponseDto
 import sopt.org.millie.data.dto.response.CategoryResponseDto
 
 interface SearchService {
     @GET("categories")
     suspend fun getCategories(): BaseResponse<List<CategoryResponseDto>>
+
+    @GET("books/{bookId}")
+    suspend fun getBookDetailInformation(
+        @Path("bookId") bookId: Long,
+    ): BaseResponse<BookDetailInformationResponseDto>
 }
